Add files via upload
A curious library that you may have never heard of is h2o. An in-memory platform for distributed and scalable machine learning, h2o can run on powerful clusters when you need boosted computing power. Additionally, h2o is a very interesting and diverse library. It contains so many diverse features (ranging from training models to automl capabilities) that it’s easy to get a bit lost when using it, particularly due to the high number of methods and functions one can use with the package. 

In this script, I am covering some examples (with code) of h2o, namely:

1- train a couple of machine learning models;
2- do some hyperparameter tuning;
3- perform an automl routine;
4- take a glimpse at the explainability module;